Replace mem_gate with mem_mlp for stronger memory gradient flow
The per-column scalar gate (sigmoid) was stuck at 0.5 — memory gradients
were 34x weaker than LM and the gate was drifting toward suppression.
Replace with a residual MLP (Linear→SiLU→Linear) that learns to combine
H_mid and memory readout, same pattern as split_mlp for PCM surprise.
Small init (std=0.01) on final layer instead of zero-init to allow
immediate gradient flow to memory graph.

+++ scripts/benchmark_v8.py
@@ -119,7 +119,7 @@ def main():
     model = V8Model(cfg).to(device).to(torch.bfloat16)
     for p_param in model.memory.parameters():
         p_param.data = p_param.data.float()
-    model.lm.mem_gate.data = model.lm.mem_gate.data.float()
+    # mem_mlp stays in model dtype (bf16)
     model.train()
 
     use_memory = not args.no_memory
